Move appnet client lib to ecs-agent
