+++ ecs-agent/api/appnet/client_linux_test.go
@@ -23,8 +23,6 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	"github.com/aws/amazon-ecs-agent/agent/api/serviceconnect"
-
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/gorilla/mux"
 	prometheus "github.com/prometheus/client_model/go"
@@ -179,7 +177,7 @@ func TestGetStats(t *testing.T) {
 			t.Cleanup(func() {
 				ts.Close()
 			})
-			stats, err := Client().GetStats(serviceconnect.RuntimeConfig{AdminSocketPath: tc.udsPath, StatsRequest: testStatsUrl, DrainRequest: testDrainUrl})
+			stats, err := Client().GetStats(tc.udsPath, testStatsUrl)
 			assert.Equal(t, tc.expectedResult, stats)
 			if tc.isErrorExpected {
 				assert.Error(t, err)
@@ -222,7 +220,7 @@ func TestDrainInboundConnections(t *testing.T) {
 			t.Cleanup(func() {
 				ts.Close()
 			})
-			err := Client().DrainInboundConnections(serviceconnect.RuntimeConfig{AdminSocketPath: tc.udsPath, StatsRequest: testStatsUrl, DrainRequest: testDrainUrl})
+			err := Client().DrainInboundConnections(tc.udsPath, testDrainUrl)
 			if tc.isErrorExpected {
 				assert.Error(t, err)
 				assert.Contains(t, err.Error(), tc.expectedErrorContains)
